requirements.txt
opencv-python
numpy
pytesseract
pyautogui
pillow
re
pynput